requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
lxml>=4.9.0
psycopg2-binary==2.9.7
